+++ src/main/java/com/dwarfeng/springterminator/node/config/SpringTerminatorDefinitionParser.java
@@ -1,12 +1,14 @@
 package com.dwarfeng.springterminator.node.config;
 
 import com.dwarfeng.springterminator.impl.handler.TerminateHandlerImpl;
+import com.dwarfeng.springterminator.sdk.util.BeanDefinitionParserUtil;
 import org.springframework.beans.factory.config.BeanDefinition;
+import org.springframework.beans.factory.config.ConstructorArgumentValues;
+import org.springframework.beans.factory.config.RuntimeBeanReference;
 import org.springframework.beans.factory.support.AbstractBeanDefinition;
 import org.springframework.beans.factory.support.BeanDefinitionBuilder;
-import org.springframework.beans.factory.xml.AbstractSingleBeanDefinitionParser;
+import org.springframework.beans.factory.xml.BeanDefinitionParser;
 import org.springframework.beans.factory.xml.ParserContext;
-import org.springframework.lang.NonNull;
 import org.springframework.util.StringUtils;
 import org.w3c.dom.Element;
 
@@ -18,42 +20,35 @@ import javax.annotation.Nonnull;
  * @author DwArFeng
  * @since 1.0.0
  */
-public class SpringTerminatorDefinitionParser extends AbstractSingleBeanDefinitionParser {
+public class SpringTerminatorDefinitionParser implements BeanDefinitionParser {
 
-    private static final String STANDARD_BEAN_NAME = "terminateHandler";
+    private static final String STANDARD_HANDLER_NAME = "terminateHandler";
+    private static final String STANDARD_CONFIG_REF = "terminateConfig";
 
     @Override
-    protected String getBeanClassName(@NonNull Element element) {
-        return TerminateHandlerImpl.class.getCanonicalName();
-    }
-
-    @Nonnull
-    @Override
-    protected String resolveId(
-            @NonNull Element element, @NonNull AbstractBeanDefinition definition, @NonNull ParserContext parserContext
-    ) {
-        if (element.hasAttribute("id") && StringUtils.hasText(element.getAttribute("id"))) {
-            return element.getAttribute("id");
+    public BeanDefinition parse(Element element, @Nonnull ParserContext parserContext) {
+        String handlerName = BeanDefinitionParserUtil.mayResolvePlaceholder(
+                parserContext, element.getAttribute("handler-name")
+        );
+        String configRef = BeanDefinitionParserUtil.mayResolvePlaceholder(
+                parserContext, element.getAttribute("config-ref")
+        );
+        if (!StringUtils.hasText(handlerName)) {
+            handlerName = STANDARD_HANDLER_NAME;
         }
-        return STANDARD_BEAN_NAME;
-    }
-
-    @Override
-    protected void doParse(
-            @NonNull Element element, @NonNull ParserContext parserContext, @NonNull BeanDefinitionBuilder builder
-    ) {
-        try {
-            if (element.hasAttribute("pre-delay")) {
-                builder.addPropertyValue("preDelay", element.getAttribute("pre-delay"));
-            }
-            if (element.hasAttribute("post-delay")) {
-                builder.addPropertyValue("postDelay", element.getAttribute("post-delay"));
-            }
-        } catch (Exception e) {
-            parserContext.getReaderContext().error("转换数字时出现异常", e);
+        if (!StringUtils.hasText(configRef)) {
+            configRef = STANDARD_CONFIG_REF;
         }
+        BeanDefinitionParserUtil.makeSureBeanNameNotDuplicated(parserContext, handlerName);
 
+        BeanDefinitionBuilder builder = BeanDefinitionBuilder.rootBeanDefinition(TerminateHandlerImpl.class);
+        builder.getRawBeanDefinition().setAutowireMode(AbstractBeanDefinition.AUTOWIRE_CONSTRUCTOR);
+        ConstructorArgumentValues constructorArgumentValues = new ConstructorArgumentValues();
+        constructorArgumentValues.addIndexedArgumentValue(1, new RuntimeBeanReference(configRef));
+        builder.getRawBeanDefinition().setConstructorArgumentValues(constructorArgumentValues);
         builder.setScope(BeanDefinition.SCOPE_SINGLETON);
         builder.setLazyInit(false);
+        parserContext.getRegistry().registerBeanDefinition(handlerName, builder.getBeanDefinition());
+        return null;
     }
 }